Simpler code paths for zero-copy and copy-out.
Pull decisions about whether to copy up out of the hot path. Now there
are two kinds of send: one that copies everything and one that copies
nothing. This makes the sends both more efficient, and more importantly
it lets us measure time more accurately. Before, we had to rdtsc inside
the hot loop because each chunk might have been copied/not-copied. The
rdtscs are now moved out of the loop, which helps since they are
expensive.

I've also tuned some parameters a bit to try to make the numbers more
representative of what we want to show. I disabled inlining; inlining
just made the performnace of zero-copy follow the performance of
copy-out for a bit since the driver was memcpy'ing the inlined data
anyway. The paper text discusses this a bit.

I've bumped the tx buffer count to 128 to be closer to what ib_send_bw
does. This does seem to boost tx bw a bit in the smaller sends,
suggesting we weren't keeping the NIC completely busy.