Changes to support Bw-tree like bench and hugepages.
Adds a final phase to the benchmarks that simulates the delta-record
style layout that Bw-tree has.

This commit also includes scripts to setup host hugepage support and to
map the simulated log with 1 GB hugepages. There are still some bugs to
work out in this part. /dev/hugetlbfs gets created but we cannot create
a file in it to mmap. I think we may need to up the nr_inodes for the
fs, but I have to investigate a bit more.

Finally, this commit hacks the chunkSizes it tests with to limit them to
the ones we actually plot in the paper. We should undo this hack, but it
also makes the experiments take much much longer.